smart_shopping_cart-main/requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
fonttools==4.55.4
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_resources==6.5.2
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.23.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==4.21.4
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
wget==3.2
zipp==3.21.0
